requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
PyJWT==2.9.0
Werkzeug==3.1.2

#pip install -r requirements.txt
